+++ cancel.go
@@ -70,5 +70,7 @@ func (api *Api) ReceiptCancel(cancelData CancelData) (APIResponse, error) {
 
 	result := APIResponse{}
 	json.NewDecoder(res.Body).Decode(&result)
+	if result == nil { result =  map[string]interface{}{} }
+	result["http_status"] = res.StatusCode
 	return result, nil
 }