+++ confirm.go
@@ -34,5 +34,7 @@ func (api *Api) ServerConfirm(receiptId string) (APIResponse, error) {
 
 	result := APIResponse{}
 	json.NewDecoder(res.Body).Decode(&result)
+	if result == nil { result =  map[string]interface{}{} }
+	result["http_status"] = res.StatusCode
 	return result, nil
 }