+++ easy.go
@@ -46,5 +46,7 @@ func (api *Api) GetUserToken(userToken EasyUserTokenPayload) (APIResponse, error
 
 	result := APIResponse{}
 	json.NewDecoder(res.Body).Decode(&result)
+	if result == nil { result =  map[string]interface{}{} }
+	result["http_status"] = res.StatusCode
 	return result, nil
 }